# pkg/v3/swagger/ca/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cas": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get All CAs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get All CAs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CACertificate"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Create CA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create CA",
                "parameters": [
                    {
                        "description": "CA Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.CreateCABody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CACertificate"
                        }
                    },
                    "400": {
                        "description": "Struct Validation error || CA type inconsistent || Issuance expiration greater than CA expiration || Incompatible expiration time ref",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cas/import": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Import CA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Import CA",
                "parameters": [
                    {
                        "description": "CA Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.ImportCABody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CACertificate"
                        }
                    },
                    "400": {
                        "description": "Struct Validation error || CA type inconsistent || Issuance expiration greater than CA expiration || Incompatible expiration time ref || CA and the provided key dont match",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cas/{id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get CA By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get CA By ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CACertificate"
                        }
                    },
                    "400": {
                        "description": "Struct Validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "CA not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Delete CA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete CA",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Struct Validation error || CA Status inconsistent",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "CA not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cas/{id}/certificates": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Get Certificates by CA",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Certificates by CA",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Certificate"
                            }
                        }
                    },
                    "400": {
                        "description": "Struct Validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "CA not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/cas/{id}/certificates/sign": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Sign Certificate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sign Certificate",
                "parameters": [
                    {
                        "description": "Sign Certificate Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.SignCertificateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Certificate"
                        }
                    },
                    "400": {
                        "description": "Struct Validation error || CA Status inconsistent",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "CA not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CACertificate": {
            "type": "object",
            "properties": {
                "certificate": {
                    "$ref": "#/definitions/models.X509Certificate"
                },
                "creation_ts": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "issuance_expiration": {
                    "$ref": "#/definitions/models.Expiration"
                },
                "issuer_metadata": {
                    "$ref": "#/definitions/models.IssuerCAMetadata"
                },
                "key_metadata": {
                    "$ref": "#/definitions/models.KeyStrengthMetadata"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "revocation_timestamp": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.CertificateStatus"
                },
                "subject": {
                    "$ref": "#/definitions/models.Subject"
                },
                "type": {
                    "$ref": "#/definitions/models.CAType"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_to": {
                    "type": "string"
                }
            }
        },
...
        "models.IssuerCAMetadata": {
            "type": "object",
            "properties": {
                "ca_name": {
                    "type": "string"
                },
                "serial_number": {
                    "type": "string"
                }
            }
        },
...
    }
}